Disable binding to thumbstick move and fix bug
When binding on a controller, it was impossible to get thumbstick button  because the thumbstick input would always trigger first.
Binding to thumbstick movement (left, up, and so on) isn't really needed imo, so it's currently disabled.
Disabling is done by checking if the key up, down, or analog is one of the thumbsticks. If yes, it handles it and moves on. If no, then it will bind to that input.
The addition of categories introduced a bug with saving and reseting inputs to default. The bug has been fixed, but optimization may be needed in the future.